backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
aiomysql
boto3
python-jose[cryptography]
passlib[bcrypt]
PyMySQL
pdfkit
weasyprint
jinja2
